Prepend resources with efs
